// frontend/src/shared/lib/api/model/blockerPayload.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chem-Vault
 * OpenAPI spec version: 0.1.0
 */
import type { BlockerPayloadSamplesItem } from './blockerPayloadSamplesItem';

export interface BlockerPayload {
  table: string;
  entity_type: string;
  fk_column: string;
  count: number;
  samples: BlockerPayloadSamplesItem[];
  truncated: boolean;
}